XEngine-V9.13.0.1001 Build20250304
其他更新
    安装脚本修正rockylinux无法清理的问题
    CI增加了新的系统发布,现在支持RockyLiunx9 ARM64了
    CI增加了新的系统发布,现在支持Windows ARM64了
核心组件
    核心安全模块
        修改代码风格,统一空格风格类型
        编解码类修改URL编解码,优化修正某些特殊字符处理问题
音视频编解码组件
    音视频帮助模块
        媒体信息类增加了获取媒体开始时间的函数
        媒体信息类获取时间修正信息流处理导致的错误
        媒体信息类获取音视频信息现在可以获取视频的B帧数和GOP大小了
系统组件
    系统函数模块
        文件类枚举文件类增加了UNICODE版本,现在支持ANSI和UNICODE函数处理
        文件类删除和创建多级目录现在支持UNICODE和ANSI版本,用于多语言处理
        文件类获取文件属性现在支持UNICODE和ANSI版本
        文件类保存缓冲区数据现在支持UNICODE和ANSI版本
        文件类创建稀疏文件函数现在支持UNICODE和ANSI版本
        文件类拷贝文件函数现在支持UNICODE和ANSI版本
基础组件
    基础模块
        字符串类分割路径函数和修复路径函数现在支持UNICODE和ANSI版本了
        字符串类删除字符串和改变字符串以及替换字符串函数现在支持UNICODE和ANSI版本了
        字符串类获取开始结束分割字符串和获取字符串KEY,VALUE函数现在支持UNICODE和ANSI版本了
        字符串类分割字符串函数现在支持UNICODE和ANSI版本了

+++ XEngine_LINEnv.sh
@@ -20,7 +20,7 @@ function InstallEnv_Print()
 	echo -e "\033[32m|***************************************************************************|\033[0m"
 	echo -e "\033[33m                 XEngine-Toolkit Linux和Mac版本环境安装脚本                    \033[0m"
 	echo -e "\033[33m                       运行环境：Linux x64 AND MacOS x64                      \033[0m"
-	echo -e "\033[33m                       脚本版本：Ver 9.12.0.1001                              \033[0m"
+	echo -e "\033[33m                       脚本版本：Ver 9.13.0.1001                              \033[0m"
 	echo -e "\033[33m                  安装环境的时候请检查所有三方库下载安装成功                     \033[0m"
 	echo -e "\033[32m|***************************************************************************|\033[0m"
 	echo -e "\033[44;37m当前时间：$m_EnvTimer 执行用户：$m_EnvExecName 你的环境：$m_EnvCurrent\033[0m"
@@ -322,13 +322,24 @@ function InstallEnv_SdkShared()
 			rm -rf /usr/local/lib/libHelpComponents_*.dylib
 			rm -rf /usr/local/lib/libRfcComponents_*.dylib
 			rm -rf /usr/local/lib/libStreamMedia_*.dylib
-		else
+		elif [ "$m_EnvRelease" -eq "2" ] ; then
 			rm -rf /usr/local/lib/libXEngine_*.so
 			rm -rf /usr/local/lib/libXClient_*.so
 			rm -rf /usr/local/lib/libNetHelp_*.so
 			rm -rf /usr/local/lib/libHelpComponents_*.so
 			rm -rf /usr/local/lib/libRfcComponents_*.so
 			rm -rf /usr/local/lib/libStreamMedia_*.so
+			rm -rf /usr/local/ffmpeg-xengine
+			ldconfig
+		fi
+		else
+			rm -rf /usr/lib64/libXEngine_*.so
+			rm -rf /usr/lib64/libXClient_*.so
+			rm -rf /usr/lib64/libNetHelp_*.so
+			rm -rf /usr/lib64/libHelpComponents_*.so
+			rm -rf /usr/lib64/libRfcComponents_*.so
+			rm -rf /usr/lib64/libStreamMedia_*.so
+			rm -rf /usr/local/ffmpeg-xengine
 			ldconfig
 		fi
 		echo -e "\033[45;37m删除共享库成功\033[0m"
